client/requirements.txt
PyJWT==1.6.4
flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-migrate==2.5.2
Flask-Script==2.0.6
